moved Hierarchical object scripts
