--- Assets/Scripts/Layers/Properties.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 1431d552360f49b28a45eeed89988c9a
-timeCreated: 1707393249
