+++ Assets/_Application/Layers/LayerTypes/HierarchicalObject.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 25bc8cacefd3f4e3bb9f6266c94fedfc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
