+++ Assets/_Application/Layers/LayerTypes/HierarchicalObject/Properties/AbstractHierarchicalObjectPropertySection.cs
@@ -1,7 +1,6 @@
-using Netherlands3D.Twin.Layers.LayerTypes;
 using UnityEngine;
 
-namespace Netherlands3D.Twin.Layers.Properties
+namespace Netherlands3D.Twin.Layers.LayerTypes.HierarchicalObject.Properties
 {
     public abstract class AbstractHierarchicalObjectPropertySection : MonoBehaviour
     {
