+++ Assets/_Application/Layers/LayerTypes/HierarchicalObject/Properties/HierarchicalObjectPropertySectionInstantiator.cs
@@ -1,7 +1,7 @@
-using Netherlands3D.Twin.Layers.LayerTypes;
+using Netherlands3D.Twin.Layers.Properties;
 using UnityEngine;
 
-namespace Netherlands3D.Twin.Layers.Properties
+namespace Netherlands3D.Twin.Layers.LayerTypes.HierarchicalObject.Properties
 {
     public class HierarchicalObjectPropertySectionInstantiator : MonoBehaviour, IPropertySectionInstantiator
     {
