+++ Assets/_Application/Layers/LayerTypes/HierarchicalObject/Properties/TransformLayerPropertyData.cs.meta
