+++ Assets/_Functionalities/AssetBundles/Scripts/AssetBundleLoader.cs
@@ -1,9 +1,8 @@
-using Netherlands3D.Twin.Layers;
 using Netherlands3D.Twin.Projects;
 using System;
 using System.Collections;
 using System.IO;
-using Netherlands3D.Twin.Layers.LayerTypes;
+using Netherlands3D.Twin.Layers.LayerTypes.HierarchicalObject;
 using UnityEngine;
 using UnityEngine.Events;
 using UnityEngine.Networking;
