+++ Assets/_Functionalities/ObjectLibrary/Scripts/ObjectLibraryButton.cs
@@ -1,7 +1,6 @@
 using System.Collections;
 using Netherlands3D.Twin;
-using Netherlands3D.Twin.Layers;
-using Netherlands3D.Twin.Layers.LayerTypes;
+using Netherlands3D.Twin.Layers.LayerTypes.HierarchicalObject;
 using UnityEngine;
 using UnityEngine.UI;
